% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\docType{data}
\name{algorithms}
\alias{algorithms}
\title{Caret Algorithms}
\format{
\subsection{\code{algorithms}}{

A \code{data.frame} with 230 rows and 60 columns:
\describe{
\item{X}{Algorithms names}
\item{Further columns}{Algorithms attributes}
}
}
}
\source{
\url{https://topepo.github.io/caret/models-clustered-by-tag-similarity.html}
}
\usage{
algorithms
}
\description{
A \code{data.frame} with characteristics of each algorithm available in \code{caretSDM}. Each
column is a different characteristics. This can be helpful for more experienced modelers select
algorithms. See the source for a selection method using this data.
}
\keyword{datasets}
