% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/band_names.R
\name{set_band_names}
\alias{set_band_names}
\alias{get_band_names}
\title{Band management for \code{stars} objects}
\usage{
set_band_names(x, new_names)

get_band_names(x)
}
\arguments{
\item{x}{A \code{stars} object.}

\item{new_names}{\code{character} vector of new names to be addressed to the bands.}
}
\value{
A \code{stars} object with updated band names.
}
\description{
Functions to retrieve names and change them from \code{stars} objects.
}
\examples{
scen
get_band_names(scen)
scen2 <- set_band_names(scen, paste0("bioclimatics_", 1:19))
get_band_names(scen2)

}
\seealso{
\code{\link{input_sdm} \link{sdm_area} \link{scen}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
