% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc.R
\docType{data}
\name{bioc}
\alias{bioc}
\title{Bioclimatic Variables}
\format{
\subsection{\code{bioc}}{

A \code{stars} with 1 attribute and 3 bands:
\describe{
\item{bio1}{Annual Mean Temperature}
\item{bio4}{Temperature Seasonality}
\item{bio12}{Annual Precipitation}
}
}
}
\source{
\url{https://www.worldclim.org/}
}
\usage{
bioc
}
\description{
A \code{stars} object with bioclimatic variables (bio1, bio4 and bio12) for the Parana state in Brazil.
Data obtained from WorldClim 2.1 at 10 arc-min resolution.
}
\keyword{datasets}
