% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialfuncs.R
\name{blockCV2fold}
\alias{blockCV2fold}
\title{Convert folds from blockCV to trainControl in caret}
\usage{
blockCV2fold(object)
}
\arguments{
\item{object}{object from spatialBlock, buffering or envBlock functions of the \pkg{blockCV} package.}
}
\value{
a list that can be passed to index and indexOut of \code{\link[caret]{trainControl}}.
}
\description{
Convert folds from blockCV to trainControl in caret
}
\note{
\code{biomod2Format} argument of the \pkg{blockCV} functions must be set to \code{FALSE}.
}
\examples{
\dontrun{
sb <- spatialBlock(speciesData = sptrain, biomod2Format = FALSE) # biomod2Format must be FALSE
indices <- blockCV2fold(sb)
control <- trainControl(index = indices$index, indexOut = indices$indexOut)

# This will result in the same output
control <- trainControlSDM(method = "none", spatial.folds = sb$foldID)
}
}
\seealso{
Other spatial: 
\code{\link{blockENM2fold}()},
\code{\link{create.test.index.blockCV}()}
}
\concept{spatial}
