% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialfuncs.R
\name{blockENM2fold}
\alias{blockENM2fold}
\title{Create Spatial blocks using ENMeval}
\usage{
blockENM2fold(
  method = "block",
  y,
  coords,
  presence.level = NULL,
  env = NULL,
  aggregation.factor = NULL
)
}
\arguments{
\item{method}{One of 'block', 'check1', 'check2'. Check ENMeval documentation}

\item{y}{A response factor}

\item{coords}{Two-column data.frame or matrix of longitude and latitude, same order as y}

\item{presence.level}{character. Value in y which represents presence. By default, the first level
of y is considered presence.}

\item{env, aggregation.factor}{used only for methods 'check1' and 'check2'}
}
\value{
A vector with folds to be used in \code{\link{trainControlSDM}}.
}
\description{
Create Spatial blocks using ENMeval
}
\seealso{
Other spatial: 
\code{\link{blockCV2fold}()},
\code{\link{create.test.index.blockCV}()}
}
\concept{spatial}
