% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_sdm.R
\name{buffer_sdm}
\alias{buffer_sdm}
\title{Create buffer around occurrences}
\usage{
buffer_sdm(occ_data, size = NULL, occ_crs = NULL, mcp = FALSE)
}
\arguments{
\item{occ_data}{A \code{data.frame} object with species, decimalLongitude and decimalLatitude columns.
Usually the output from \code{GBIF_data}.}

\item{size}{\code{numeric}. The distance between the record and the margin of the buffer (i.e.
buffer radius).}

\item{occ_crs}{\code{numeric}. Indicates which EPSG is the \code{occ_data} in.}

\item{mcp}{\code{boolean}. Should the buffer be applied in each record (FALSE) or in a minimum
convex polygon/convex hull (TRUE)? Standard is \code{FALSE}.}
}
\value{
A \code{sf} buffer around \code{occ_data} records.
}
\description{
Create buffer around records in \code{occ_data} to be used as study area
}
\examples{
# Create sdm_area object:
study_area <- buffer_sdm(occ, size=50000, occ_crs=6933)
plot(study_area)

}
\seealso{
\code{\link{GBIF_data}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
