% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretSDM-package.R
\docType{package}
\name{caretSDM-package}
\alias{caretSDM}
\alias{caretSDM-package}
\title{caretSDM: Build Species Distribution Modeling using 'caret'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Use machine learning algorithms and advanced geographic information system tools to build Species Distribution Modeling in a extensible and modern fashion.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://luizesser.github.io/caretSDM/}
  \item Report bugs at \url{https://github.com/luizesser/caretSDM/issues}
}

}
\author{
\strong{Maintainer}: Luíz Fernando Esser \email{luizesser@gmail.com} (\href{https://orcid.org/0000-0003-2982-7223}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Luíz Fernando Esser \email{luizesser@gmail.com} (\href{https://orcid.org/0000-0003-2982-7223}{ORCID}) [copyright holder]
  \item Reginaldo Ré (\href{https://orcid.org/0000-0001-6452-3466}{ORCID})
  \item Marcos R. Lima (\href{https://orcid.org/0000-0002-5901-0911}{ORCID})
  \item Edivando Couto (\href{https://orcid.org/0000-0003-4264-8449}{ORCID})
  \item José Hilário Delconte Ferreira (\href{https://orcid.org/0000-0002-7116-2600}{ORCID})
  \item Valéria Batista (\href{https://orcid.org/0000-0002-6574-7338}{ORCID})
  \item Dayani Bailly (\href{https://orcid.org/0000-0002-6954-9902}{ORCID})
}

}
\keyword{internal}
