% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ci_95}
\alias{ci_95}
\alias{se}
\title{Functions to calculate confidence intervals}
\usage{
ci_95(x, alpha = 0.05, na.rm = FALSE)

se(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector to calculate confidence intervals}

\item{alpha}{Probability used for the Student t Distribution
(\code{alpha = 0.05} is equivilant 95\% CI)}

\item{na.rm}{a logical value indicating whether NA values should be stripped.}
}
\description{
\code{ci_95} calculates 95\% confidence intervals and \code{se} calculates stardard error.
}
