% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict2.R
\name{clamp_data}
\alias{clamp_data}
\title{Clamp values in a dataset}
\usage{
clamp_data(object, dataset)
}
\arguments{
\item{object}{A data.frame with the training data or a model returned
by \code{\link[caret]{train}}, from which the training data is obtained.}

\item{dataset}{Either a data.frame or a rasterStack object with values to be clamped.
Columns or layers names should match the columns of the training data.}
}
\value{
The \code{dataset} with values clamped.
}
\description{
This function clamp values in the \code{dataset} object based on
minimum and maximum values on the train data.
}
