% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R, R/importance.R, R/jackknife.R,
%   R/misc.R, R/plotROC.R, R/response.R, R/utils.R
\name{evaluate.list}
\alias{evaluate.list}
\alias{c.evaluate.train}
\alias{varImp2.list}
\alias{c.varImp2}
\alias{jackknife.list}
\alias{c.jackknife.train}
\alias{confusionMatrix2.list}
\alias{ROCcurve.list}
\alias{c.ROCcurve}
\alias{c.response.train}
\alias{combine}
\title{Combine objects}
\usage{
\method{evaluate}{list}(x, ...)

\method{c}{evaluate.train}(...)

\method{varImp2}{list}(model, ...)

\method{c}{varImp2}(...)

\method{jackknife}{list}(model, ...)

\method{c}{jackknife.train}(...)

\method{confusionMatrix2}{list}(model, ...)

\method{ROCcurve}{list}(model, ...)

\method{c}{ROCcurve}(...)

\method{c}{response.train}(...)
}
\arguments{
\item{x, model}{A list of models created by \code{\link[caret]{train}}.}

\item{...}{For \code{function(model.list, ...)}, further arguments passed to the function.
For \code{c(...)}, multiple objects returned by the same function.}
}
\description{
You can use \code{function(model.list)} or \code{c(...)}
to combine multiple objects generated by functions from this package.
}
\details{
The output \code{function(list(model1, model2))} is the same as \code{c(function(model1), function(model2))}.
But note that you can only concatenate objects obtained from models with different methods (e.g. 'rf','gbm', ...).
\cr\cr
In contrast, using \code{function(list(model1, model2))} will append '.1', '.2', etc
if a method is repeated in the models.
Optionally, you can furnish a named list, and the names will override the labels of the methods.
}
