% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{confusionMatrix2}
\alias{confusionMatrix2}
\alias{confusionMatrix2.train}
\title{Confusion Matrix for a caret model}
\usage{
confusionMatrix2(model, ...)

\method{confusionMatrix2}{train}(model, testdata = NULL, testy = NULL, ...)
}
\arguments{
\item{model}{A model returned by \code{\link[caret]{train}}.}

\item{...}{Further arguments to \code{\link[caret]{confusionMatrix}}.}

\item{testdata}{A data.frame with test values to be evaluated.
If \code{NULL}, training values will be used.}

\item{testy}{A factor of response variable of \code{testdata}.
If \code{NULL}, it will be guessed from \code{testdata} data.frame.}
}
\description{
This is a wrapper function for \code{\link[caret]{confusionMatrix}},
but it accepts directly a caret model and a data.frame with test data.
}
\note{
Using \code{confusionMatrix2(model)} (without test data) will produce simple training metrics, unlike
\code{confusionMatrix(model)}, which is based on resampling procedure.
}
\seealso{
\code{\link{evaluate}} \code{\link{ROCcurve}}
}
