% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialfuncs.R
\name{create.test.index.blockCV}
\alias{create.test.index.blockCV}
\title{Create resampling index for test data using blockCV}
\usage{
create.test.index.blockCV(speciesData, sb)
}
\arguments{
\item{speciesData}{A spatialPoints or spatialPointsDataFrame with the locations for each test data.}

\item{sb}{A \code{spatialBlock} object.}
}
\value{
A list of test data index to be used in each resampling.
}
\description{
This function provides resamples for test data to be used in \code{\link{evaluate}} or \code{\link{ROCcurve}}
The resamples are based in the blocks created by \code{\link[blockCV]{spatialBlock}}.
}
\details{
The folds of each block are assigned to the test data based on their location.
If a point does not fall over any block, then the fold of the block closest to that point is
assigned, with a warning.\cr\cr
For each resample, all folds except one are used, and this is done for each fold.\cr\cr
The order of points in the spatialPoints object must be the same from the data.frame used for the test data.
}
\seealso{
Other spatial: 
\code{\link{blockCV2fold}()},
\code{\link{blockENM2fold}()}
}
\concept{spatial}
