% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{create.test.index}
\alias{create.test.index}
\title{Create resampling index for test data}
\usage{
create.test.index(
  testy,
  control = NULL,
  method = "boot",
  number = ifelse(grepl("cv", method), 10, 25),
  repeats = 10,
  p = 0.75
)
}
\arguments{
\item{testy}{A factor of response variable of the test data.}

\item{control}{A \code{trainControl} object or a \code{train} model from which a trainControl object is used.}

\item{method, number, repeats, p}{Check \code{\link[caret]{trainControl}} for details. Only used if \code{control = NULL}.}
}
\value{
A list of test data index to be used in each resampling.
}
\description{
This function provides resamples for test data to be used in \code{\link{evaluate}} or \code{\link{ROCcurve}}
Arguments are similar to \code{\link[caret]{trainControl}}.
}
\details{
When \code{control} is provided, arguments "method", "number", "repeats" and "p" are obtained from the \code{trainControl} 
object and used to create the index for test data. When \code{control = NULL}, the arguments provided
in this function are used instead.
}
\seealso{
\code{\link{create.test.index.blockCV}}
}
