% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{createEnsemble}
\alias{createEnsemble}
\alias{predict.ensemble.train}
\title{Create a ensemble from models}
\usage{
createEnsemble(
  model.list,
  ensemble_method = "weighted_mean",
  metric = NULL,
  calc.pred = FALSE
)

\method{predict}{ensemble.train}(object, newdata = NULL, type = "raw", ...)
}
\arguments{
\item{model.list}{A list of models returned by \code{\link[caret]{train}}.}

\item{ensemble_method}{One of 'mean', 'median', 'weighted_mean' and 'number_votes'. Check details.}

\item{metric}{A metric to use to calculate weights. Only used if \code{ensemble_method = "weighted.mean"}.
Must be one of the metrics provided by 'summaryFunction' used to train the models.
If \code{NULL}, the metric of the first model is used.}

\item{calc.pred}{logical. Calculate predictions for training data? If \code{TRUE}, savePredictions is set
to "final"; else, savePredictions is set to "FALSE".}

\item{object}{A object returned by \code{createEnsemble}.}

\item{newdata}{A data.frame containing data to predict.}

\item{type}{One of "raw" or "prob".}

\item{...}{Further arguments passed to \code{predict}.}
}
\value{
An S3 object of classes "ensemble.train" that also inherits "train".
This object can be used in other functions, like \code{evaluate} or \code{confidence_map}.
}
\description{
Combine the predictions of multiple models
}
\details{
You can create a ensemble model based on predictions of multiple models.
The ensemble prediction is calculated based on the ensemble_method:
\itemize{
\item \code{mean} - models mean prediction.
\item \code{median} - models median prediction.
\item \code{weighted_mean } - models weighted mean prediction. Weighs are based on the \code{metric},
so models with higher metric have more weight in the mean.
\item \code{number_votes} - The number of predictions of the first class (considered positive or presence) is
divided by the number of models.  When the prediction is close to 1, it means that all models
agree to predict the first class. Only models of type "Classification" are supported.
}
For classification, the probability of predictions is used to create the ensemble.
}
\seealso{
\code{\link{confidence_map}}.
}
