% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_columns.R
\name{.find_columns}
\alias{.find_columns}
\title{Find species, longitude and latitudecolumns in a data.frame.}
\usage{
.find_columns(df, col_names = NULL, spp = TRUE)
}
\arguments{
\item{df}{data.frame to search for.}

\item{col_names}{Names of the columns so it doesn't need to look for.}

\item{spp}{Boolean. Search for species column?}
}
\value{
Vector of column names that correspond to the species, longitude and latitude columns.
}
\description{
Find species, longitude and latitudecolumns in a data.frame.
}
\keyword{internal}
