% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_to_df_sdm.R
\name{.sf_to_df_sdm}
\alias{.sf_to_df_sdm}
\title{Transform a occurrences simplefeature in a data.frame.}
\usage{
.sf_to_df_sdm(occ_sf)
}
\arguments{
\item{occ_sf}{SimpleFeatures to transform in data.frame. Usually from occurrences slot.}
}
\value{
Data.frame with three columns: "species", "decimalLongitude" and "decimalLatitude".
}
\description{
Transform a occurrences simplefeature in a data.frame.
}
\keyword{internal}
