% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_input_sdm}
\alias{is_input_sdm}
\alias{is_sdm_area}
\alias{is_occurrences}
\alias{is_models}
\alias{is_predictions}
\title{\code{is_class} functions to check caretSDM data classes.}
\usage{
is_input_sdm(x)

is_sdm_area(x)

is_occurrences(x)

is_models(x)

is_predictions(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\value{
Boolean.
}
\description{
This functions returns a boolean to check caretSDM object classes.
}
\examples{
# Create sdm_area object:
sa <- sdm_area(parana, cell_size = 25000, output_crs = 6933)

is_sdm_area(sa)

is_input_sdm(sa)

}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
