% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{make_grid}
\alias{make_grid}
\title{Make grid}
\usage{
make_grid(
  shp,
  cell_width = 0,
  cell_height = 0,
  var_names = NULL,
  centroid = F,
  epsg = NULL
)
}
\arguments{
\item{pred}{A predictors object}

\item{th}{Threshold}

\item{maxobservations}{Number of sets of pseudoabsence to create}
}
\value{
A predictors object with VIF data
}
\description{
Build a grid out of given shape and variables.
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
