% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent.predict.R
\name{methods.maxent}
\alias{methods.maxent}
\alias{predict.maxent}
\alias{summary.maxent}
\alias{plot.maxent}
\alias{var_imp.maxent}
\title{Methods for maxent object}
\usage{
\method{predict}{maxent}(
  object,
  newdata = NULL,
  thrtype = NULL,
  clamp = NULL,
  type = "raw",
  ...
)

\method{summary}{maxent}(object, ...)

\method{plot}{maxent}(x, testdata = NULL, testy = NULL, ...)

\method{var_imp}{maxent}(object, itype = "both", ...)
}
\arguments{
\item{newdata}{A data.frame of values to predict. If \code{NULL}, predictions of the training samples are returned.}

\item{thrtype}{Which thr type to use. Only used if \code{type = 'raw'}.}

\item{clamp}{logical. Apply clamping in prediction?}

\item{type}{One of 'raw' or 'prob'.}

\item{...}{Further arguments to \code{\link[pROC]{plot.roc}}.}

\item{x, object}{A model returned by \code{\link{maxent}}.}

\item{testdata}{A data.frame with test values to be use in ROC curve. If \code{NULL}, training values will be used.}

\item{testy}{A factor of response variable of \code{testdata}. If \code{NULL}, it will be guessed from \code{testdata} data.frame.}

\item{itype}{One of 'contribution', 'permutation' or 'both'.}
}
\description{
predict(), summary(), plot() and var_imp() methods for maxent object
}
\details{
\itemize{
\item predict - return predicted classes or matrix of class probabilities
\item summary - open report generated by maxent.jar
\item plot - Plot a ROC curve based in the model
\item var_imp - Get variable importance for a maxent model
}
}
