% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{n_records}
\alias{n_records}
\title{Miscelaneous functions}
\usage{
n_records(x)

predictors_names(x)

scenarios_names(x)
}
\arguments{
\item{x}{A input_sdm object.}
}
\description{
Set of functions to obtain
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
