% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.input_data}
\alias{plot.input_data}
\title{S3 Methods for plot}
\usage{
plot.input_data(
  x,
  what = NULL,
  spp_name = NULL,
  scenario = NULL,
  id = NULL,
  pa = TRUE,
  ensemble_type = "mean_occ_prob"
)
}
\arguments{
\item{x}{Object to be plotted.}
}
\value{
A predictors object.
}
\description{
This function creates different plots depending on the input.
}
\seealso{
\code{\link{WorldClim_data}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
