% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.occurrences}
\alias{plot.occurrences}
\title{S3 Methods for plot}
\usage{
fun(arg1, arg2 = default, ...)
}
\arguments{
\item{x}{Object to be plotted.}
}
\value{
A predictors object.
}
\description{
This function creates different plots depending on the input.
}
\seealso{
\code{\link{WorldClim_data}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
