% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict2.R
\name{predict2}
\alias{predict2}
\alias{predict2.train}
\alias{predict2.list}
\alias{predict2.RasterStack}
\title{Extract predictions train objects using a threshold}
\usage{
predict2(object, ...)

\method{predict2}{train}(object, newdata = NULL, type = "raw", ...)

\method{predict2}{list}(object, ...)

\method{predict2}{RasterStack}(object, model, ...)
}
\arguments{
\item{object, model}{A model returned by \code{\link[caret]{train}} or a list of models.
object can also be a RasterStack from the \code{raster} package, used to create predictions.}

\item{...}{Further arguments passed to \code{predict} or method 'train'.}

\item{newdata}{An optional set of data to predict on.}

\item{type}{One of "raw", "prob", "both" or "prob_presence" (see details).}
}
\value{
A vector or data.frame with predictions or probabilities, based on the argument \code{type}. Check details.
If \code{model} is a list of models, than a list of predictions is returned.
}
\description{
This function is a wrapper around \code{\link[caret]{predict.train}}, but it accepts a 
threshold cutoff to calculate predictions from class probabilities.
}
\details{
Most arguments work the same as \code{\link[caret]{predict.train}}.\cr\cr
This function is only useful for two class Classification problems, when using 
\code{type = "raw"} and a different threshold from the default (which must set using \code{\link{setThreshold}}).
Otherwise, the output is the same as \code{predict}.\cr\cr
Additionally, you can set \code{type = "both"}, which will return a data.frame with both class probabilities and a prediction.
or \code{type = "prob1"} to return a single vector with the first class probabilities (assumed to be presence).\cr\cr
The method for RasterStack is a simpler version of \code{raster::predict}, but it may be faster sometimes.
}
\seealso{
\code{\link{setThreshold}}
}
