% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_misc.R
\name{predictors_names}
\alias{predictors_names}
\title{Add predictors to sdm_area}
\usage{
predictors_names(sdm_area)
}
\arguments{
\item{x}{A \code{sf}, \code{RasterStack}, \code{SpatRaster} or \code{stars}.}

\item{cell_size}{Numeric. The cell size to be used in models.}

\item{epsg}{Numeric. If epsg from \code{x} is missing,}
}
\value{
A \code{sdm_area} object.
}
\description{
This function adds a new predictor to the sdm_area object.
}
\seealso{
\code{\link{WorldClim_data}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
