% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_sdm.R
\name{print.models}
\alias{print.models}
\title{Print method for models}
\usage{
\method{print}{models}(x, ...)
}
\arguments{
\item{x}{models object}

\item{...}{passed to other methods}
}
\description{
Print method for models
}
