% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrences_sdm.R
\name{print.occurrences}
\alias{print.occurrences}
\title{Print method for occurrences}
\usage{
\method{print}{occurrences}(x, ...)
}
\arguments{
\item{x}{occurrences object}

\item{...}{passed to other methods}
}
\value{
Concatenate structured characters to showcase what is stored in the object.
}
\description{
Print method for occurrences
}
