% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_sdm.R
\name{print.predictions}
\alias{print.predictions}
\title{Print method for predictions}
\usage{
\method{print}{predictions}(x, ...)
}
\arguments{
\item{x}{predictions object}

\item{...}{passed to other methods}
}
\value{
Concatenate structured characters to showcase what is stored in the object.
}
\description{
Print method for predictions
}
