% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_scenarios.R, R/sdm_area.R
\name{print.sdm_area}
\alias{print.sdm_area}
\title{Print method for predictors}
\usage{
\method{print}{sdm_area}(x)

\method{print}{sdm_area}(x)
}
\description{
Print method for predictors

Print method for predictors
}
