% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R, R/maxent.predict.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{evaluate}
\alias{dotplot}
\alias{pairs}
\alias{var_imp}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{evaluate}}, \code{\link[generics]{var_imp}}}

  \item{graphics}{\code{\link[graphics]{pairs}}}

  \item{lattice}{\code{\link[lattice:xyplot]{dotplot}}}
}}

