% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotROC.R
\name{ROCcurve}
\alias{ROCcurve}
\alias{ROCcurve.train}
\alias{plot.ROCcurve}
\title{Plot a ROC curve}
\usage{
ROCcurve(model, ...)

\method{ROCcurve}{train}(
  model,
  testdata = NULL,
  testy = NULL,
  testindex = NULL,
  thrs = NULL,
  use.pROC = FALSE,
  calc.train = TRUE,
  errorFunction = ci_95,
  ...
)

\method{plot}{ROCcurve}(
  x,
  type = "facet",
  select.thr = "all",
  plot_ci = TRUE,
  add.text = TRUE,
  ...
)
}
\arguments{
\item{model}{A model or a list of models returned by \code{\link[caret]{train}}.}

\item{...}{Arguments passed to \code{predict}. Don't use object, newdata, type.}

\item{testdata}{A data.frame with test values to be use in ROC curve. If \code{NULL}, training values will be used.}

\item{testy}{A factor of response variable of \code{testdata}.
If \code{NULL}, it will be guessed from \code{testdata} data.frame.}

\item{testindex}{A list with rows index of testdata for each resample, preferably an output of
\code{\link{create.test.index}} or \code{\link{create.test.index.blockCV}}.
If \code{NULL}, indexes are create based on
the same methods provided by \code{trainControl} used in the model. Ignored if \code{use.pROC = TRUE}.}

\item{thrs}{A named vector with thresholds values between [0,1] to be plotted in the ROC curve
or the output of \code{\link{thresholder2}}. If \code{model} is a list, this should be also a list of the same ordet
with thresholds to be applied for each model.}

\item{use.pROC}{logical. If \code{TRUE}, use \code{\link[pROC]{roc}}
to compute the ROC curve and confidence intervals.
If \code{FALSE}, use "model" and "testindex" resamples instead.}

\item{calc.train}{logical. If \code{FALSE}, ROC curve for train data is not computed.}

\item{errorFunction}{A function used to calculate confidence interval across resamples.
Function is ignored if \code{use.pROC = TRUE}. Else, the  ROC curve is calculated from the
mean between resamples. If \code{NULL}, confidence interval is not calculated and all train and
test data are used to calculate the curve. Default is 95\% confidence interval.}

\item{x}{An objected returned by ROCcurve.}

\item{type}{one of "facet", "train" or "test".}

\item{select.thr}{"all", "none" or a vector with indices of thresholds to plot.}

\item{plot_ci}{logical. Plot confidence intervals?}

\item{add.text}{logical. Plot threshold methods names?}
}
\value{
An S3 object of class 'ROCcurve', including:
\item{dat.roc }{A data.table with values of Sensitivity and Specificity to plot the ROC curve.}
\item{thrs}{A data.table with Sensitivity and Specificity for each threshold provided by \code{thrs}.}
\item{coords}{A data.table with the coordinates to plot the confidence interval.}
}
\description{
Plot a ROC curve from a caret model.
}
\examples{
\dontrun{
# using pROC package
ROCcurve(model, use.pROC = TRUE)

# using caret resamples, with testdata and thresholds
testindex <- create.test.index(testy)
thrs <- thresholder2(model)
r.obj <- ROCcurve(model, testdata, testy, testindex, thrs, use.pROC = FALSE)

plot(r.obj, type = "facet", select.thr = 3:7, plot_ci = TRUE, add.text = FALSE)

# for multiple models
models <- list(model1, model2, model3)
thrs <- lapply(models, thresholder2)
r.obj <- ROCcurve(models, thrs, testdata, testindex=testindex)
plot(r.obj, type = "test", select.thr = 3:7, plot_ci = FALSE, add.text = TRUE)
}
}
\seealso{
\code{\link{evaluate}} \code{\link{confusionMatrix2}}
}
