% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salm.R
\docType{data}
\name{salm}
\alias{salm}
\title{Salminus brasiliensis occurrence data}
\format{
\subsection{\code{salm}}{

A \code{data.frame} with 46 rows and 3 columns (EPSG:6933):
\describe{
\item{species}{Species name}
\item{decimalLongitude}{Longitude in meters}
\item{decimalLatitude}{Latitude in meters}
}
}
}
\source{
\url{https://www.gbif.org}
}
\usage{
salm
}
\description{
A \code{data.frame} object with Salminus brasiliensis occurrence data obtained from GBIF and
filtered with Parana state \code{sf}.
}
\keyword{datasets}
