% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios_sdm.R
\name{scenarios_sdm}
\alias{scenarios_sdm}
\alias{scenarios_names}
\alias{get_scenarios_data}
\title{Scenarios Managing}
\usage{
scenarios_sdm(x,
              study_area = NULL,
              vars_study_area = NULL,
              predictors_names = NULL,
              rescaling = NULL,
              scenarios_names = NULL)

scenarios_names(i)

get_scenarios_data(i)
}
\arguments{
\item{x}{A \code{character} vector with the path to stack files, a \code{stars}, a
\code{RasterStack}, a \code{SpatRaster} or a \code{list}.}

\item{study_area}{A \code{sf} object with the study area to mask scenarios. If \code{NULL}, the
whole area is used.}

\item{vars_study_area}{Logical. Should the function use the table within the \code{study_area} as
predictors? Standard is \code{NULL}.}

\item{predictors_names}{A \code{character} vector with names to be addressed to each predictor.}

\item{rescaling}{A list of parameters to pass on rescaling function (optional, see details).}

\item{scenarios_names}{A \code{character} vector with names to be addressed to each scenario.}

\item{i}{\code{input_sdm} object.}
}
\value{
A \code{scenarios} object.
}
\description{
This function creates and manage \code{scenarios} objects.
}
\details{
If using a \code{sdm_area} object as predictors, add scenarios through \code{add_scenarios}
function.

Rescaling is particularly usefull for aquatic environments, once creates a grid around riverlines
allowing modelers to account for the surrounding environment. Rescaling needs a \code{list} with
\code{cell_size} and \code{epsg}.

\code{scenarios_names} returns the scenarios names in \code{input_sdm} object.

\code{get_scenarios_data} retrieve scenarios data as a \code{list} of \code{sf}s to each scenario.
}
\examples{
i <- input_sdm(
     occurrences_sdm(occ_data),
     predictors_sdm(bioc, study_area = parana),
     scenarios_sdm(scen, study_area = parana)
     )
i

}
\seealso{
\code{\link{input_sdm} \link{sdm_area} \link{add_scenarios} \link{scen}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
