% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholder2.R
\name{thresholder2}
\alias{thresholder2}
\alias{plot.thresholder2}
\alias{summary.thresholder2}
\alias{as.data.frame.thresholder2}
\title{Choose a Probability Threshold}
\usage{
thresholder2(
  model,
  thr.method = "all",
  thr.interval = 0.01,
  final = TRUE,
  add.statistics = "J",
  obs.prev = NULL,
  FPC = 1,
  FNC = 1,
  req.sens = 0.85,
  req.spec = 0.85
)

\method{plot}{thresholder2}(
  x,
  select.thr = "all",
  statistics = "Sensitivity+Specificity",
  display.final = TRUE,
  add.text = TRUE,
  plot = TRUE,
  ...
)

\method{summary}{thresholder2}(
  object,
  which.statistics = "J",
  maximize = ifelse(which.statistics == "Dist", FALSE, TRUE),
  ...
)

\method{as.data.frame}{thresholder2}(x, ...)
}
\arguments{
\item{model}{A model returned by \code{\link[caret]{train}}.}

\item{thr.method}{A vector with threshold methods to use when calculating the optimal threshold.
Can be a character or numeric vector. Check details for possible values. Default "all" is to use all methods.}

\item{thr.interval}{A value to create probability thresholds cutoffs intervals to be evaluated. Should be between 0 and 0.1.}

\item{final}{logical. Should only the final tuning parameters chosen by train be used?}

\item{add.statistics}{A character vector indicating additional statistics to calculate. See \code{\link[caret]{thresholder}} for a list of possible values.
Note that 'Sensitivity', 'Specificity', 'Kappa', 'Dist' and 'Detection Prevalence' are always calculated.}

\item{obs.prev}{Observed prevalence, in case your data is taken from a larger dataset. Defaults to observed prevalence from training data.}

\item{FPC, FNC}{False Positive and False Negative Costs. Used only to calculate "Cost" threshold.}

\item{req.sens, req.spec}{Required Sensitivity and Specificity to calculate "ReqSens" and "ReqSpec" thresholds.}

\item{x, object}{An object returned by thresholder2}

\item{select.thr}{A vector with threshold methods to be displayed. Can be index of 
thr.method used in thresholder2.}

\item{statistics}{A character vector indicating statistics to be displayed. Can be index of 
statistics used in thresholder2. Alternatively, can be "all" to display all available statistics
or Sensitivity+Specificity" to display a Sensitivity and Specificity in a single plot.}

\item{display.final}{logical. Should display only thresholds for the final model tuned by \code{train}?
This is only used if you set \code{final = FALSE} in \code{thresholder2}.}

\item{add.text}{logical. Plot threshold methods names? They are displayed only if \code{display.final = TRUE}.}

\item{plot}{logical. Return the plot or not?}

\item{...}{ignored}

\item{which.statistics}{A single character with a statistic to be used to select the best threshold.}

\item{maximize}{A logical: should the statistic be maximized or minimized?}
}
\value{
An S3 object of class 'thresholder2', including:
\itemize{
  \item thrs - A data.table with thresholds methods and their statistics.
  \item thresholder - The output of \code{\link[caret]{thresholder}}.
  }

For the method \code{plot}, if \code{plot = TRUE}, a ggplot is returned. If \code{plot = FALSE}, a list of data.tables
in the long format (used by ggplot) is returned.
}
\description{
This function apply \code{thresholder} and return the optimal thresholds based on multiple methods (check details).
}
\details{
Possible \code{thr.method} are:
\itemize{
 \item Default - The default for most models, this set the threshold to 0.5
 \item Min_Presence - Required threshold to have Sensitivity = 1
 \item 10%_Presence - Required threshold to have Sensitivity = 0.9
 \item Sens=Spec - Optimize threshold where sensitivity equals specificity
 \item MaxSens+Spec - Optimize threshold which maximaze sensitivity plus specificity
 \item MaxKappa - Optimize threshold which maximaze kappa
 \item PredPrev=Obs - Optimize threshold where Observed prevalence equals specificity
 \item Dist - Optimize threshold which minimize Dist (check \code{\link[caret]{thresholder}})
 \item Cost - Optimize threshold which maximize Cost (check \code{PresenceAbsence::optimal.thresholds})
 \item ReqSens - Required threshold to have Sensitivity = \code{req.sens}
 \item ReqSpec - Required threshold to have Specificity = \code{req.spec}
}
You can check the help documentation of \code{PresenceAbsence::optimal.thresholds} or \code{caret::thresholder2}
for more details. Most names of threshold methods are the same for \code{optimal.thresholds}.
}
\note{
By default, Youden's J statistic is also returned, which is the same as TSS (True Skill Statistic).
}
\examples{
\dontrun{
# Select a threshold looking at all tuning parameters
t.obj <- thresholder2(model, thr.method= 1:6, final = FALSE)
plot(t.obj, select.thr="all", statistics="all", display.final = FALSE)
summary(t.obj)

# Select a threshold looking only at the final model
t.obj <- thresholder2(model, thr.interval = 0.005, final = TRUE)
plot(t.obj, select.thr=3:7)
thr <- summary(t.obj)
thr # this will hold the best threshold based on which.statistics

as.data.frame(t.rf) # data.frame with thresholds and statistics for each method
}
}
\seealso{
\code{\link[caret]{thresholder}}
}
