% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{twoClassSDM}
\alias{twoClassSDM}
\alias{twoClassSDM2}
\alias{twoClassTSS}
\title{Calculates performance metrics for two classes}
\usage{
twoClassSDM(data, lev = NULL, model = NULL)

twoClassSDM2(data, lev = NULL, model = NULL)

twoClassTSS(data, lev = NULL, model = NULL)
}
\arguments{
\item{data, lev, model}{Check \code{\link[caret]{twoClassSummary}}.}
}
\description{
Functions to calculate performance metrics for two class Classification problems,
to be used in \code{trainControl} or \code{evaluate}.
}
\details{
\code{twoClassTSS} calculates ROC, sensitivity, specificity and True-Skill Statistics (TSS).
\code{twoClassSDM} is a wrapper for multiple metrics, and calculates kappa, AUC and TSS.
\code{twoClassSDM2} calculates all metrics from \code{twoClassSDM}, plus log-likelihood.
}
\examples{
\dontrun{
control <- trainControl(method = "cv", classProbs = TRUE, summaryFunction = twoClassSDM)
control <- trainControl(method = "cv", summaryFunction = twoClassTSS)
}
}
