% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_predictions.R
\name{write_predictions}
\alias{write_predictions}
\title{Write Predictions}
\usage{
write_predictions(x, l)
}
\arguments{
\item{m}{A models object}

\item{scen}{A scenarios object}

\item{th}{Thresholds for metrics to be used}
}
\value{
A predictions object
}
\description{
This function writes predictions in the given directory.
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
