% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_predictors.R
\name{add_predictors}
\alias{add_predictors}
\title{Add predictors to sdm_area}
\usage{
add_predictors(sdm_area, pred)
}
\arguments{
\item{x}{A shapefile or a raster.}

\item{cell_size}{Numeric. The cell size to be used in models.}

\item{epsg}{Numeric. If epsg from x is missing,}
}
\value{
A sdm_area object.
}
\description{
This function adds a new predictor to the sdm_area object
}
\seealso{
\code{\link{WorldClim_data}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
