% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_scenarios.R
\name{add_scenarios}
\alias{add_scenarios}
\title{Include scenarios in the SDM workflow.}
\usage{
add_scenarios(sdm_area, scen = NULL, ...)
}
\arguments{
\item{sdm_area}{A sdm_area object.}

\item{scen}{Scenarios to be added: RasterStack, stars, SpatRaster or a directory. If NULL adds predictors as a scenario.}
}
\value{
A sdm_area object.
}
\description{
This function includes scenarios in the sdm_area object.
}
\seealso{
\code{\link{WorldClim_data}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
