% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_scenarios.R
\name{add_scenarios}
\alias{add_scenarios}
\title{Add scenarios to \code{sdm_area}}
\usage{
add_scenarios(sa, scen = NULL, scenarios_names = NULL, pred_as_scen = TRUE,
                     variables_selected = NULL, stationary = NULL, ...)
}
\arguments{
\item{sa}{A \code{sdm_area} object.}

\item{scen}{\code{RasterStack}, \code{SpatRaster} or \code{stars} object. If \code{NULL} adds
predictors as a scenario.}

\item{variables_selected}{Character vector with variables names in \code{scen} to be used as
predictors. If \code{NULL} adds all variables.}

\item{stationary}{Names of variables from \code{sa} that should be used in scenarios as
stationary variables.}
}
\value{
The input \code{sdm_area} object with a new slot called scenarios with \code{scen} data
as a \code{list}, where each slot of the \code{list} is a scenario and each scenario is a
\code{sf}.
}
\description{
This function includes scenarios in the \code{sdm_area} object.
}
\examples{
# Create sdm_area object:
sa <- sdm_area(parana, cell_size = 25000, crs = 6933)

# Include predictors:
sa <- add_predictors(sa, bioc) |> dplyr::select(c("bio01", "bio12"))

# Include scenarios:
sa <- add_scenarios(sa)

}
\seealso{
\code{\link{sdm_area} \link{input_sdm}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
