% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{background}
\alias{background}
\alias{n_background}
\alias{background_data}
\title{Obtain Background data}
\usage{
background(occ,
           pred = NULL,
           n = 10000,
           n_set = 1,
           proportion = NULL)

n_background(i)

background_data(i)
}
\arguments{
\item{occ}{A \code{occurrences_sdm} or \code{input_sdm} object.}

\item{pred}{A \code{sdm_area} object. If \code{NULL} and \code{occ} is a \code{input_sdm},
\code{pred} will be retrieved from \code{occ}.}

\item{n}{\code{numeric}. Number of background records to be generated in each dataset created.
If \code{NULL} then the function prevents imbalance by using the same number of presence records
(\code{n_records(occ)}). If you want to address different sizes to each species, you must provide
a named vector (as in \code{n_records(occ)}).}

\item{n_set}{\code{numeric}. Number of datasets of background data to create.}

\item{proportion}{\code{numeric}. A number between 0 and 1 representing a proportion of the area
to be mapped as background. E.g.: if the whole area has 5,000 cells and proportion is 0.1, then
\code{n} is set to 500. Standard is NULL. This argument overwrites \code{n}.}

\item{i}{A \code{input_sdm} object.}
}
\value{
A \code{occurrences_sdm} or \code{input_sdm} object with background data.
}
\description{
This function obtains background data given a set of predictors.
}
\details{
\code{background} is used in the SDM workflow to obtain background data, a step necessary for
MaxEnt algorithm to run. This function helps avoid the use of pseudoabsence data in background
algorithms and the use of background data in pseudoabsence algorithms, a very common mistake.

\code{n_background} returns the number of background records obtained per species.

\code{background_data} returns a \code{list} of species names. Each species name will have a
\code{list}s with background data from class \code{sf}.
}
\examples{
# Create sdm_area object:
sa <- sdm_area(parana, cell_size = 25000, crs = 6933)

# Include predictors:
sa <- add_predictors(sa, bioc) |> select_predictors(c("bio1", "bio4", "bio12"))

# Include scenarios:
sa <- add_scenarios(sa)

# Create occurrences:
oc <- occurrences_sdm(occ, crs = 6933) |> join_area(sa)

# Create input_sdm:
i <- input_sdm(oc, sa)

# Pseudoabsence generation:
i <- background(i, proportion = 1) # All available data is obtained as background data.

}
\seealso{
\code{link{input_sdm} \link{pseudoabsences} \link{occurrences_sdm}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
