% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc.R
\docType{data}
\name{bioc}
\alias{bioc}
\title{Bioclimatic Variables}
\format{
## `bioc`
A \code{stars} with 1 attribute and 2 bands:
\describe{
  \item{bio01}{Annual Mean Temperature}
  \item{bio12}{Annual Precipitation}
}
}
\source{
<https://www.worldclim.org/>
}
\usage{
bioc
}
\description{
A \code{stars} object with bioclimatic variables (bio01 and bio12) for the Parana state in Brazil.
Data obtained from WorldClim 2.1 at 10 arc-min resolution.
}
\keyword{datasets}
