% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretSDM-package.R
\docType{package}
\name{caretSDM-package}
\alias{caretSDM}
\alias{caretSDM-package}
\title{caretSDM: A package for building Species Distribution Modeling in R using caret.}
\description{
This package is intended to help modelers to build Species Distribution Modeling (SDMs) using a machine learning approach. The package uses caret as a underlying package to build and project models.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/luizesser/caretSDM/issues}
}

}
\author{
\strong{Maintainer}: Luíz Esser \email{luizesser@gmail.com} (\href{https://orcid.org/0000}{ORCID})

Authors:
\itemize{
  \item Dayani Baili (\href{https://orcid.org/0000}{ORCID})
  \item Edivando Couto (\href{https://orcid.org/0000}{ORCID})
  \item José Hilário Delconte Ferreira (\href{https://orcid.org/0000}{ORCID})
  \item Reginaldo Ré (\href{https://orcid.org/0000}{ORCID})
  \item Valéria Batista (\href{https://orcid.org/0000}{ORCID})
}

}
\keyword{internal}
