% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{data_clean}
\alias{data_clean}
\title{Presence data cleaning wrapper}
\usage{
data.clean(x, r = NULL)
}
\arguments{
\item{occ}{A occurrences object}

\item{pred}{A predictors object}

\item{terrestrial}{If the species is terrestrial (TRUE), than the function deletes non-terrestrial coordinates.}
}
\description{
Data cleaning procedure using CoordinateCleaner and a raster.
}
\details{
Function to clean presence data avoiding autocorrelation.
}
\examples{
# Select species names:
s <- c("Araucaria angustifolia", "Paubrasilia echinata", "Eugenia uniflora")

# Run function:
data <- GBIF_data(s)

# Clean coordinates:
data2 <- data.clean(data)

}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
