% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{data_clean}
\alias{data_clean}
\title{Presence data cleaning routine}
\usage{
data_clean(x,
           pred = NULL,
           species = NA,
           lon = NA,
           lat = NA,
           terrestrial = TRUE,
           independent_test = TRUE)
}
\arguments{
\item{occ}{A \code{occurrences} object or \code{input_sdm}.}

\item{pred}{A \code{predictors} object. If \code{occ} is a \code{input_sdm} object with
predictors data, than \code{pred} is obtained from it.}

\item{species}{A \code{character} stating the name of the column with species names in \code{occ} (see details).}

\item{lon}{A \code{character} stating the name of the column with longitude in \code{occ} (see details).}

\item{lat}{A \code{character} stating the name of the column with latitude in \code{occ} (see details).}

\item{terrestrial}{If the species is terrestrial (TRUE), than the function deletes
non-terrestrial coordinates.}

\item{independent_test}{TRUE. If \code{occ} has independent test data, the data cleaning routine
is also applied on it.}
}
\value{
A \code{occurrences} object or \code{input_sdm} with new presence data.
}
\description{
Data cleaning wrapper using CoordinateCleaner.
}
\details{
If the user does not used \code{GBIF_data} function to obtain species records, the function may
have problems to find which column from the presences table has species, longitude and latitude
information. In this regard, we implemented the parameters \code{species}, \code{lon} and
\code{lat} so the use can explicitly inform which columns should be used. If they remain as NA
(standard) the function will try to guess which columns are the correct one.
}
\examples{
# Create sdm_area object:
sa <- sdm_area(parana, cell_size = 25000, crs = 6933)

# Include predictors:
sa <- add_predictors(sa, bioc) |> dplyr::select(c("bio01", "bio12"))

# Include scenarios:
sa <- add_scenarios(sa)

# Create occurrences:
oc <- occurrences_sdm(occ, crs = 6933) |> join_area(sa)

# Create input_sdm:
i <- input_sdm(oc, sa)

# Clean coordinates:
i <- data_clean(i)

}
\seealso{
\code{\link{GBIF_data} \link{occurrences} \link{sdm_area} \link{input_sdm}
\link{predictors}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
