% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_sdm.R
\name{ensemble_sdm}
\alias{ensemble_sdm}
\alias{get_ensembles}
\alias{add_ensembles}
\title{Predict SDM models in new data}
\usage{
ensemble_sdm(m,
            scen = NULL,
            method = "average",
            metric = NULL,
            fun = NULL
            )

get_ensembles(i)

add_ensembles(e1, e2)
}
\arguments{
\item{m}{A \code{input_sdm} or a \code{models} object.}

\item{scen}{A \code{scenarios} object or \code{NULL}. If \code{NULL} and \code{m} is a
\code{input_sdm} with a scenarios slot, it will be used.}

\item{method}{Character or a function. Which ensembles should be calculated? See details.}

\item{metric}{Character. Used with \code{method = "weighted_average"}: Which metric should
be used to weight predictions? If NULL}

\item{fun}{Function. If \code{method = "committee_average"}, the function will be used to binarize
the data. It will receive caret's train object and must return a numeric value (the threshold,
see details).}

\item{i}{A \code{input_sdm} or a \code{predictions} object.}

\item{e1}{A \code{ensembles} object.}

\item{e2}{A \code{ensembles} object.}
}
\value{
A \code{input_sdm} or a \code{predictions} object.
}
\description{
This function projects SDM models to new scenarios
}
\details{
\code{ensembles} could be set to three different strategies OR a custom function.
The three implemented strategies are:
\code{average} is the mean occurrence probability, which is a simple mean of predictions;
\code{weighted_average} is the same average, but weighted by a metric, which needs to be
set using argument \code{metric} (see \link{mean_validation_metrics} for the metrics available).
\code{committee_average} is the committee average, as known as majority rule, where predictions
are binarized and then a mean is obtained. To binarize predictions, user can set a custom function
in the \code{fun} argument to calculate a threshold for each model. Standardly, the
committee average uses the \code{caret::thresholder} function to find the threshold that
maximizes the sum of sensitivity and specificity (through \code{caretSDM:::.MaxSeSp}). The custom
function must use the argument \code{mod}, which is the model output from caret package (see
\code{\link{get_models}}) and must return a \code{numeric} value (see example).

\code{get_predictions} returns the \code{list} of all predictions to all scenarios, all species,
all algorithms and all repetitions. Useful for those who wish to implement their own ensemble
methods.

\code{get_ensembles} returns a \code{matrix} of \code{data.frame}s, where each column is a
scenario and each row is a species.

\code{scenarios_names} returns the scenarios names in a \code{sdm_area} or \code{input_sdm}
object.

\code{get_scenarios_data} returns the data from scenarios in a \code{sdm_area} or
\code{input_sdm} object.
}
\examples{
if (interactive()) {
  # Create sdm_area object:
  set.seed(1)
  sa <- sdm_area(parana, cell_size = 100000, crs = 6933)

  # Include predictors:
  sa <- add_predictors(sa, bioc) |> select_predictors(c("bio1", "bio12"))

  # Include scenarios:
  sa <- add_scenarios(sa)

  # Create occurrences:
  oc <- occurrences_sdm(occ, crs = 6933) |> join_area(sa)

  # Create input_sdm:
  i <- input_sdm(oc, sa)

  # Pseudoabsence generation:
  i <- pseudoabsences(i, method="random", n_set=2)

  # Custom trainControl:
  ctrl_sdm <- caret::trainControl(method = "boot",
                                  number = 1,
                                  repeats = 1,
                                  classProbs = TRUE,
                                  returnResamp = "all",
                                  summaryFunction = summary_sdm,
                                  savePredictions = "all")

  # Train models:
  i <- train_sdm(i, algo = c("naive_bayes"), ctrl=ctrl_sdm) |>
    suppressWarnings()

  # Predict models:
  i  <- predict_sdm(i, th = 0.8)

  # Ensemble:
  i <- ensemble_sdm(i, method = "average")
  i
}

# Example from a custom function to obtain the threshold that maximizes
# the sensitivity plus specificity:
MaxSeSp <- function(mod) {
   th <- caret::thresholder(mod,
                            threshold = seq(0, 1, by = 0.001),
                            final = TRUE,
                            statistics = c("Sensitivity", "Specificity")
                            )
   th <- th$prob_threshold[which.max(th$Sensitivity + th$Specificity)]
   if (length(th) > 1) mean(th) else th
}

}
\seealso{
\code{\link{sdm_area} \link{input_sdm} \link{mean_validation_metrics}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
