% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIF_data.R
\name{GBIF_data}
\alias{GBIF_data}
\title{Retrieve Species data from GBIF}
\usage{
GBIF_data(s)
}
\arguments{
\item{s}{vector of species names.}
}
\description{
Calculate the area of a species weighting it by species occurrence or environmental suitability.
}
\details{
Function to retrieve species data from GBIF.
}
\examples{
# Select species names:
s <- c("Araucaria angustifolia", "Paubrasilia echinata", "Eugenia uniflora")

# Run function:
data <- GBIF_data(s)
}
\references{
https://www.gbif.org
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
