% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIF_data.R
\name{GBIF_data}
\alias{GBIF_data}
\title{Retrieve Species data from GBIF}
\usage{
GBIF_data(s)
}
\arguments{
\item{s}{\code{character} vector of species names.}

\item{file}{File to save the output.}

\item{...}{Arguments to pass on \code{rgbif::occ_data()}.}
}
\description{
This function is a wrapper to get records from GBIF using \code{rgbif} and return a
\code{data.frame} ready to be used in caretSDM.
}
\examples{
# Select species names:
s <- c("Araucaria angustifolia", "Paubrasilia echinata", "Eugenia uniflora")

# Run function:
data <- GBIF_data(s)

}
\references{
https://www.gbif.org
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
