% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcms_ensembles.R
\name{gcms_ensembles}
\alias{gcms_ensembles}
\title{Ensemble GCMs into one scenario}
\usage{
gcms_ensembles(i, gcms = NULL)
}
\arguments{
\item{i}{A \code{input_sdm} object.}

\item{gcms}{GCM codes in \code{scenarios_names(i)} to group scenarios.}
}
\value{
A \code{input_sdm} object with grouped GCMs.
}
\description{
An ensembling method to group different GCMs into one SSP scenario
}
\examples{
# Create sdm_area object:
sa <- sdm_area(parana, cell_size = 25000, crs = 6933)

# Include predictors:
sa <- add_predictors(sa, bioc) |> dplyr::select(c("bio01", "bio12"))

# Include scenarios:
sa <- add_scenarios(sa)

# Create occurrences:
oc <- occurrences_sdm(occ, crs = 6933) |> join_area(sa)

# Create input_sdm:
i <- input_sdm(oc, sa)

# Clean coordinates:
i <- data_clean(i)

}
\seealso{
\code{\link{GBIF_data} \link{occurrences_sdm} \link{sdm_area} \link{input_sdm}
\link{predictors}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
