% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_sdm.R
\name{input_sdm}
\alias{input_sdm}
\title{Create a input_sdm object}
\usage{
input_sdm(...)
}
\arguments{
\item{...}{Data to be used in SDMs. Can be a occurrences, a predictors and/or a scenarios object.}
}
\value{
A input_sdm object.
}
\description{
This function creates a new predictors object
}
\seealso{
\code{\link{occurrences}\link{predictors}\link{scenarios}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
