% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_sdm.R
\name{input_sdm}
\alias{input_sdm}
\title{\code{input_sdm}}
\usage{
input_sdm(...)
}
\arguments{
\item{...}{Data to be used in SDMs. Can be a \code{occurrences}, a \code{predictors}, a
\code{scenarios}, and/or a \code{sdm_area} object. Ideally, \code{occurrences} and
\code{sdm_area} are used.}
}
\value{
A \code{input_sdm} object containing:
   \item{grid}{\code{sf} with \code{POLYGON} geometry representing the grid for the study area.}
   \item{bbox}{Four corners for the bounding box (class \code{bbox}): minimum value of X, minimum value of Y, maximum value of X, maximum value of Y}
   \item{cell_size}{\code{numeric} information regarding the size of the cell used to rescale variables to the study area, representing also the cell size in the \code{grid}.}
   \item{epsg}{\code{character} information about the EPSG used in all slots from \code{sdm_area}.}
   \item{predictors}{\code{character} vector with predictors names included in \code{sdm_area}.}
}
\description{
This function creates a new \code{input_sdm} object.
}
\details{
If \code{sdm_area} is used, it can include predictors and scenarios. In this case,
\code{input_sdm} will detect and include as \code{scenarios} and \code{predictors} in the
\code{input_sdm} output. Objects can be included in any order, since the function will work by
detecting their classes.
The returned object is used throughout the whole workflow to apply functions.
}
\examples{
# Create sdm_area object
sa <- sdm_area(parana, cell_size = 25000, epsg = 6933)

# Include predictors
sa <- add_predictors(sa, bioc)

i <- input_sdm(occurrences(occ), sa)

}
\seealso{
\code{\link{occurrences} \link{predictors} \link{scenarios} \link{sdm_area}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
