% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_input_sdm}
\alias{is_input_sdm}
\title{is_class functions to check caretSDM data classes.}
\usage{
is_input_sdm(x)
}
\arguments{
\item{x}{caretSDM object.}
}
\value{
Boolean.
}
\description{
This functions returns a boolean to check caretSDM object classes.
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
