% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicollinearity_sdm.R
\name{multicollinearity_sdm}
\alias{multicollinearity_sdm}
\alias{selected_variables}
\title{Multicollinearity Analysis}
\usage{
multicollinearity_sdm(pred,
                             method = NULL,
                             variables_selected = NULL,
                             cumulative_proportion = 0.99,
                             th = 0.5,
                             ...)

selected_variables(i)
}
\arguments{
\item{pred}{A \code{input_sdm} or \code{predictors} object.}

\item{method}{Which method should be used to detect multicollinearity. Can be a \code{character}
or a custom \code{function}.}

\item{variables_selected}{A vector with pre-selected variables names to filter variables.}

\item{cumulative_proportion}{A \code{numeric} with the threshold for cumulative proportion in
PCA. Standard is 0.99, meaning that axes returned as predictors sum up more than 99% of
environmental variance.}

\item{th}{Threshold to be applied in VIF routine. See ?usdm::vifcor.}

\item{...}{Further arguments to be passed to the applied method.}

\item{i}{A \code{input_sdm} object.}
}
\value{
A \code{input_sdm} or \code{predictors} object with VIF data.
}
\description{
Apply multicollinearity calculation on predictors.
}
\details{
multicollinearity_sdm is a wrapper function to run usdm::vifcor, usdm::vifstep or a pca
in caretSDM, but also provides a way to implement custom functions to reduce multicollinearity.
If user provides a custom function, it must have the arguments \code{env_sf} and \code{occ_sf},
which will consist of two \code{sf}s. The first has the predictor values for the whole study
area, while the second has the presence records for the species. The function must return a
vector with selected variables.
}
\examples{
# Create sdm_area object:
sa <- sdm_area(parana, cell_size = 25000, crs = 6933)

# Include predictors:
sa <- add_predictors(sa, bioc) |> select_predictors(c("bio1", "bio4", "bio12"))

# Include scenarios:
sa <- add_scenarios(sa, scen)

# Create occurrences:
oc <- occurrences_sdm(occ, crs = 6933)

# Create input_sdm:
i <- input_sdm(oc, sa)

# VIF calculation:
i <- multicollinearity_sdm(i, method = "vifcor", th = 0.5)
i

# Retrieve information about vif:
vif_summary(i)
selected_variables(i)

# Example of custom function:
custom_function <- function(env_sf, occ_sf) {
  env_df <- dplyr::select(sf::st_drop_geometry(env_sf), -"cell_id")
  correlations <- cor(env_df)
  col <- caret::findCorrelation(correlations, cutoff = 0.7)
  selected <- colnames(correlations)[-col]
  return(selected)
}

}
\seealso{
\code{\link{vif_predictors} \link{pca_predictors} \link{get_predictor_names}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
