% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ.R
\docType{data}
\name{occ}
\alias{occ}
\title{Araucaria angustifolia occurrence data}
\format{
## `occ`
A \code{data.frame} with 420 rows and 3 columns:
\describe{
  \item{species}{Species name}
  \item{decimalLongitude}{Longitude in meters}
  \item{decimalLatitude}{Latitude in meters}
}
}
\source{
<https://www.gbif.org>
}
\usage{
occ
}
\description{
A \code{data.frame} object with Araucaria angustifolia occurrence data obtained from GBIF and
filtered with Parana state \code{sf}.
}
\keyword{datasets}
