% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrences.R
\name{occurrences}
\alias{occurrences}
\title{Create a occurrences object}
\usage{
occurrences(x, ...)
}
\arguments{
\item{x}{A data.frame}

\item{...}{A vector with column names addressing the columns with species names, longitude and latitude.}
}
\value{
A occurrences object.
}
\description{
This function creates a occurrences object
}
\examples{
df <- data.frame(spp_names = rep("Aa", 100), longitude = runif(100), decimalLatitude = runif(100))
occ <- occurrences(df)

}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
