% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_predictors.R
\name{pca_predictors}
\alias{pca_predictors}
\title{Calculate PCA}
\usage{
pca_predictors(pred, variables_selected = NULL)
}
\arguments{
\item{pred}{A predictors object}

\item{variables_selected}{Subset of variables to be included in the PCA}
}
\value{
A predictors object with PCA data
}
\description{
Apply PCA calculation in a predictors object.
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
