% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdp_sdm.R
\name{pdp_sdm}
\alias{pdp_sdm}
\alias{get_pdp_sdm}
\title{Model Response to Variables}
\usage{
pdp_sdm(models, variables_selected = NULL)

get_pdp_sdm(i, spp = NULL, algo = NULL, variables_selected = NULL)
}
\arguments{
\item{i}{A \code{input_sdm} object.}

\item{spp}{A \code{character} vector with species names to obtain the PDPs. If \code{NULL}
(standard), the first species in \code{species_names(i)} is used.}

\item{algo}{A \code{character} containing the algorithm to obtain the PDP. If \code{NULL}
(standard) all algorithms are mixed.}

\item{variables_selected}{If there is a subset of predictors that should be ploted in this, it
can be informed using this parameter.}
}
\value{
A plot (for \code{pdp_sdm}) or a data.frame (for \code{get_pdp_sdm}) with PDP values.
}
\description{
Obtain the Partial Dependence Plots (PDP) to each variable.
}
\details{
If
}
\examples{
# Create sdm_area object:
sa <- sdm_area(parana, cell_size = 25000, epsg = 6933)

# Include predictors:
sa <- add_predictors(sa, bioc)

# Create input_sdm:
i <- input_sdm(occurrences(occ), sa)

# Clean coordinates:
i <- data_clean(i)

# VIF calculation:
i <- vif_predictors(i)
i

# Pseudoabsence generation:
i <- pseudoabsence(i, method="bioclim", variables_selected = "vif")

# Train models:
i <- train_sdm(i, algo = c("nnet", "kknn"), variables_selected = "vif")
i

# PDP plots:
pdp_sdm(i)
get_pdp_sdm(i)

}
\seealso{
\code{\link{varImp_sdm}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
