% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_occurrences}
\alias{plot_occurrences}
\alias{plot_predictors}
\alias{plot_predictions}
\title{S3 Methods for plot}
\usage{
plot_occurrences(i, spp_name = NULL, pa = TRUE)

plot_predictors(i, variables_selected = NULL)

plot_predictions(
  i,
  spp_name = NULL,
  scenario = NULL,
  id = NULL,
  ensemble = TRUE,
  ensemble_type = "mean_occ_prob"
)
}
\arguments{
\item{i}{Object to be plotted.}
}
\value{
A predictors object.
}
\description{
This function creates different plots depending on the input.
}
\seealso{
\code{\link{WorldClim_data}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
