% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_sdm.R
\name{predict_sdm}
\alias{predict_sdm}
\title{Predict SDM models in new data}
\usage{
predict_sdm(m, scen, th = 0.8, tp = "prob", file = NULL)
}
\arguments{
\item{m}{A models object}

\item{scen}{A scenarios object}

\item{th}{Thresholds for metrics to be used}
}
\value{
A models object
}
\description{
This function projects SDM models to new scenarios
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
