% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{predictors}
\alias{predictors}
\alias{predictors.sdm_area}
\alias{set_predictor_names}
\alias{set_predictor_names.sdm_area}
\alias{get_predictor_names}
\alias{get_predictor_names.sdm_area}
\title{Predictors Names Managing}
\usage{
predictors(x)

\method{predictors}{sdm_area}(x, ...)

set_predictor_names(x, new_names)

\method{set_predictor_names}{sdm_area}(x, new_names)

get_predictor_names(x)

\method{get_predictor_names}{sdm_area}(x)
}
\arguments{
\item{x}{A \code{sdm_area} object to get/set predictors names.}

\item{new_names}{\code{character} vector from size \code{length(get_predictor_names(x))}}
}
\value{
A \code{character} vector with predictors names.
}
\description{
This function manage predictors names in \code{sdm_area} objects.
}
\details{
This functions is available so users can modify predictors names to better represent them. Use
carefully to avoid giving wrong names to the predictors. Useful to make sure the predictors names
are equal the names in scenarios.
}
\examples{
sa <- sdm_area(parana, cell_size = 25000, crs = 6933)
predictors(sa)

}
\seealso{
\code{\link{parana} \link{sdm_area}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
