% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors.R
\name{predictors}
\alias{predictors}
\title{Create a predictors object}
\usage{
predictors(x, ...)
}
\arguments{
\item{x}{A string with the path to GCM files.}

\item{predictors_names}{A vector with names to be addressed to each predictor.}
}
\value{
A predictors object.
}
\description{
This function creates a new predictors object
}
\seealso{
\code{\link{WorldClim_data}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
