% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoabsences.R
\name{pseudoabsences}
\alias{pseudoabsences}
\alias{n_pseudoabsences}
\alias{pseudoabsence_method}
\alias{pseudoabsence_data}
\title{Obtain Pseudoabsences}
\usage{
pseudoabsences(occ,
               pred = NULL,
               method = "random",
               n_set = 10,
               n_pa = NULL,
               variables_selected = NULL,
               th = 0,
               size = 1,
               crs = 4326,
               mcp = FALSE)

n_pseudoabsences(i)

pseudoabsence_method(i)

pseudoabsence_data(i)
}
\arguments{
\item{occ}{A \code{occurrences_sdm} or \code{input_sdm} object.}

\item{pred}{A \code{sdm_area} object. If \code{NULL} and \code{occ} is a \code{input_sdm},
\code{pred} will be retrieved from \code{occ}.}

\item{method}{Method to create pseudoabsences. One of: "random", "bioclim", "mahal.dist" or
"buffer_sdm". User can also provide a custom function (see details).}

\item{n_set}{\code{numeric}. Number of datasets of pseudoabsence to create.}

\item{n_pa}{\code{numeric}. Number of pseudoabsences to be generated in each dataset created.
If \code{NULL} then the function prevents imbalance by using the same number of presence records
(\code{n_records(occ)}). If you want to address different sizes to each species, you must provide
a named vector (as in \code{n_records(occ)}).}

\item{variables_selected}{A vector with variables names to be used while building pseudoabsences.
Only used when method is not "random".}

\item{th}{\code{numeric} Threshold to be applied in bioclim/mahal.dist projections. See details.}

\item{size}{\code{numeric} The distance between the record and the margin of the buffer (i.e.
buffer radius).}

\item{crs}{\code{numeric} Indicates which EPSG it the size in.}

\item{mcp}{\code{boolean}. Should the buffer be applied in each record (FALSE) or in a minimum
convex polygon/convex hull (TRUE)? Standard is \code{FALSE}.}

\item{i}{A \code{input_sdm} object.}
}
\value{
A \code{occurrences_sdm} or \code{input_sdm} object with pseudoabsence data.
}
\description{
This function obtains pseudoabsences given a set of predictors.
}
\details{
\code{pseudoabsences} is used in the SDM workflow to obtain pseudoabsences, a step necessary for
most of the algorithms to run. We implemented four methods: \code{"random"}, which is
self-explanatory, \code{"buffer_sdm"}, \code{"mahal.dist"} and \code{"bioclim"}. The two last are
built with the idea that pseudoabsences should be environmentally different from presences. Thus,
we implemented two presence-only methods to infer the distribution of the species. \code{"bioclim"}
uses an envelope approach (bioclimatic envelope), while \code{"mahal.dist"} uses a distance
approach (mahalanobis distance). \code{th} parameter enters here as a threshold to binarize those
results. Pseudoabsences are retrieved outside the projected distribution of the species. If user
provides a custom function, it must have the arguments \code{env_sf} and \code{occ_sf}, which will
consist of two \code{"sf"}s. The first has the predictor values for the whole study area, while
the second has the presence records for the species. The function must return a vector with
cell_ids of the pseudoabsences, which is the first column of both objects. For \code{buffer_sdm},
user needs to specifiy the size of the buffer compatible with buffer CRS.

\code{n_pseudoabsences} returns the number of pseudoabsences obtained per species.

\code{pseudoabsence_method} returns the method used to obtain pseudoabsences.

\code{pseudoabsence_data} returns a \code{list} of species names. Each species name will have a
\code{list}s with pseudoabsences data from class \code{sf}.
}
\examples{
# Create sdm_area object:
sa <- sdm_area(parana, cell_size = 25000, crs = 6933)

# Include predictors:
sa <- add_predictors(sa, bioc) |> select_predictors(c("bio1", "bio4", "bio12"))

# Include scenarios:
sa <- add_scenarios(sa)

# Create occurrences:
oc <- occurrences_sdm(occ[1:50,], crs = 6933)

# Create input_sdm:
i <- input_sdm(oc, sa)

# Pseudoabsence generation:
i0 <- pseudoabsences(i, method="random")

# Custom method example:
buffer_pa_custom <- function(env_sf, occ_sf, buffer_dist = 3) {
  # Create buffer around occurrence points
  buffer <- sf::st_buffer(occ_sf, dist = buffer_dist)

  # Union buffers into a single geometry
  buffer_union <- sf::st_union(buffer)

  # Identify cells outside the buffer
  outside_buffer <- sf::st_difference(env_sf, buffer_union)[,1]

  # Randomly extract cell_ids outside the buffer
  pa_ids_sample <- sample(outside_buffer$cell_id, nrow(occ_sf))

  return(pa_ids_sample)
}

i1 <- pseudoabsences(i, method=buffer_pa_custom)

}
\seealso{
\code{link{input_sdm} \link{background} \link{occurrences_sdm} \link{get_occurrences}
\link{get_predictors} }
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
