% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoabsences.R
\name{pseudoabsences}
\alias{pseudoabsences}
\title{Indlude pseudoabsence data in a occurrences object}
\usage{
pseudoabsences(
  occ,
  pred = NULL,
  method = "random",
  n_set = 10,
  n_pa = NULL,
  variables_selected = NULL
)
}
\arguments{
\item{occ}{A occurrences object}

\item{pred}{A predictors object}

\item{method}{Method to create pseudoabsences}

\item{n_set}{Number of sets of pseudoabsence to create}

\item{n_pa}{Number of pseudoabsences to be generated in each dataset (vector of size number of species, with species names)}
}
\value{
A occurrences object with pseudoabsence data
}
\description{
This function includes pseudoabsences in a given occurrences object given a set of predictors
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
