% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_area.R
\name{sdm_area}
\alias{sdm_area}
\title{Create a sdm_area object}
\usage{
sdm_area(x, cell_size, epsg)
}
\arguments{
\item{x}{A shapefile or a raster.}

\item{cell_size}{Numeric. The cell size to be used in models.}

\item{epsg}{Numeric. If epsg from x is missing,}
}
\value{
A sdm_area object.
}
\description{
This function creates a new sdm_area object.
}
\examples{
# Create sdm_area object
sa <- sdm_area(parana, cell_size = 25000, epsg = 6933)

}
\seealso{
\code{\link{WorldClim_data}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
