% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_area.R
\name{sdm_area}
\alias{sdm_area}
\title{Create a \code{sdm_area} object}
\usage{
sdm_area(x, cell_size, epsg)
}
\arguments{
\item{x}{A shapefile or a raster. Usually a shapefile from \code{sf} class, but rasters from
\code{stars}, \code{rasterStack} or \code{SpatRaster} class are also allowed.}

\item{cell_size}{\code{numeric}. The cell size to be used in models.}

\item{epsg}{\code{numeric}. Indicates which EPSG should the output grid be in. If \code{NULL},
epsg from \code{x} is used.}
}
\value{
A \code{sdm_area} object.
}
\description{
This function creates a new \code{sdm_area} object.
}
\details{
The function returns a \code{sdm_area} object with a grid built upon the \code{x} parameter.
}
\examples{
# Create sdm_area object
sa <- sdm_area(parana, cell_size = 25000, epsg = 6933)

}
\seealso{
\code{\link{WorldClim_data} \link{parana} \link{input_sdm}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
