% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_as.R
\name{sdm_as_stars}
\alias{sdm_as_stars}
\title{sdm_as_X functions to transform caretSDM data into other classes.}
\usage{
sdm_as_stars(x, what = NULL, spp = NULL, scen = NULL, id = NULL, ens = NULL)
}
\arguments{
\item{x}{caretSDM object}

\item{what}{Sometimes multiple objects inside x could be transformed. This parameter allows users to specify what needs to be converted.}
}
\value{
The output is the desired class
}
\description{
This functions transform data in a caretSDM object to be used in other packages.
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
