% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{select_predictors}
\alias{select_predictors}
\alias{select.sdm_area}
\alias{select.input_sdm}
\alias{mutate.sdm_area}
\alias{mutate.input_sdm}
\alias{filter.sdm_area}
\alias{filter.input_sdm}
\alias{filter.occurrences}
\alias{filter_species}
\title{Tidyverse methods for caretSDM objects}
\usage{
select_predictors(x, ...)

\method{select}{sdm_area}(x, ...)

\method{select}{input_sdm}(x, ...)

\method{mutate}{sdm_area}(x, ...)

\method{mutate}{input_sdm}(x, ...)

\method{filter}{sdm_area}(x, ...)

\method{filter}{input_sdm}(x, ...)

\method{filter}{occurrences}(x, ...)

filter_species(x, spp = NULL, ...)
}
\arguments{
\item{x}{\code{sdm_area} object}

\item{...}{\code{character} vector with predictors to be selected.}
}
\description{
Set of functions to facilitate the use of caretSDM through tidyverse grammatics.
}
\examples{
# Create sdm_area object:
sa <- sdm_area(parana, cell_size = 25000, crs = 6933)

# Include predictors:
sa <- add_predictors(sa, bioc) |> dplyr::select(c("bio01", "bio12"))

}
