% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{select.sdm_area}
\alias{select.sdm_area}
\alias{select.input_sdm}
\title{Tidyverse methods for caretSDM objects}
\usage{
select(x, ...)

\method{select}{input_sdm}(x, ...)
}
\arguments{
\item{x}{\code{sdm_area} object}

\item{...}{\code{character} vector with predictors to be selected.}
}
\description{
Set of functions to facilitate the use of caretSDM through tidyverse grammatics.
}
\examples{
# Create sdm_area object
sa <- sdm_area(parana, cell_size = 25000, crs = 6933)

# Include predictors
sa <- add_predictors(sa, bioc)

# Select predictors
predictors(sa)
sa <- select(sa, c("bio01", "bio12"))

}
