% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_sdm.R
\name{train_sdm}
\alias{train_sdm}
\title{Train SDM models}
\usage{
train_sdm(occ, pred, algo, ctrl = NULL)
}
\arguments{
\item{occ}{A occurrences object}

\item{pred}{A predictors object}

\item{algo}{Algorithms to be used}
}
\value{
A models object
}
\description{
This function fit SDM models
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
