% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne_sdm.R
\name{tsne_sdm}
\alias{tsne_sdm}
\title{tSNE}
\usage{
tsne_sdm(occ, pred = NULL, variables_selected = NULL)
}
\arguments{
\item{occ}{A \code{occurrences} or \code{input_sdm} object.}

\item{pred}{A \code{predictors} object. If \code{occ} is of class \code{input_sdm}, then
\code{pred} is retrieved from it.}

\item{selected_vars}{Variable to be used in t-SNE. It can also be 'vif', if previously calculated.}
}
\value{
A plot to each pseudoabsence dataset in the form of a list.
}
\description{
This function calculates tSNE with presences and pseudoabsences data and returns a list of plots.
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
