% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_mem.R
\name{use_mem}
\alias{use_mem}
\title{MacroEcological Models (MEM) in caretSDM}
\usage{
use_mem(x, add = TRUE, name = "MEM")
}
\arguments{
\item{x}{a occurrences object or input_sdm containing occurrences.}

\item{add}{Logical. Should the new MEM records be added to the pool (TRUE) os species or the output should have only the summed records (FALSE)? Standard is TRUE.}

\item{name}{How should the new records be named? Standard is "MEM".}
}
\value{
A input_sdm or occurrences object with a sum o
}
\description{
This functions sums all species records into one. Should be used before the data cleaning routine.
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
