% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImp_sdm.R
\name{varImp_sdm}
\alias{varImp_sdm}
\title{Calculation of variable importance for models}
\usage{
varImp_sdm(m, id = NULL, ...)
}
\arguments{
\item{m}{A \code{models} or \code{input_sdm} object.}

\item{id}{Vector of model ids to filter varImp calculation.}

\item{...}{Parameters passing to caret::varImp().}
}
\value{
A \code{data.frame} with variable importance data.
}
\description{
This function retrieves variable importance as a function of ROC curves to each predictor.
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
