% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif_predictors.R
\name{vif_predictors}
\alias{vif_predictors}
\title{Calculate VIF}
\usage{
vif_predictors(
  pred,
  th = 0.5,
  maxobservations = 5000,
  variables_selected = NULL
)
}
\arguments{
\item{pred}{A predictors object}

\item{th}{Threshold}

\item{maxobservations}{Number of sets of pseudoabsence to create}
}
\value{
A predictors object with VIF data
}
\description{
Apply VIF calculation in a predictors object.
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
