% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif_predictors.R
\name{vif_predictors}
\alias{vif_predictors}
\title{Calculate VIF}
\usage{
vif_predictors(pred, th = 0.5, maxobservations = 5000, variables_selected = NULL)
}
\arguments{
\item{pred}{A \code{input_sdm} or \code{predictors} object.}

\item{th}{Threshold to be applied in VIF routine.}

\item{maxobservations}{Max observations to use to calculate the VIF.}

\item{variables_selected}{If there is a subset of predictors that should be used in this
function, it can be informed using this parameter.}
}
\value{
A \code{input_sdm} or \code{predictors} object with VIF data.
}
\description{
Apply Variance Inflation Factor (VIF) calculation.
}
\examples{
# Create sdm_area object:
sa <- sdm_area(parana, cell_size = 25000, crs = 6933)

# Include predictors:
sa <- add_predictors(sa, bioc) |> dplyr::select(c("bio01", "bio12"))

# Include scenarios:
sa <- add_scenarios(sa)

# Create occurrences:
oc <- occurrences_sdm(occ, crs = 6933) |> join_area(sa)

# Create input_sdm:
i <- input_sdm(oc, sa)

# Clean coordinates:
i <- data_clean(i)

# VIF calculation:
i <- vif_predictors(i)

}
\seealso{
\code{\link{usdm::vifcor}\link{predictors_names}}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
