% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_ensembles}
\alias{write_ensembles}
\alias{write_predictions}
\alias{write_models}
\alias{write_gpkg}
\alias{write_gpkg.sdm_area}
\alias{write_occurrences}
\alias{write_grid}
\title{Write caretSDM data}
\usage{
write_ensembles(x, path = "results/ensembles", ext = ".tif")

write_predictions(x, path = "results/predictions", ext = ".tif")

write_models(x, path = "results/models")

write_gpkg(x, file_path, file_name)

\method{write_gpkg}{sdm_area}(x, file_path, file_name)

write_occurrences(x, path = "results/occurrences.csv", grid = FALSE, ...)

write_grid(x, path = "results/grid_study_area.gpkg", ...)
}
\arguments{
\item{x}{Object to be written. Can be of class \code{input_sdm}, \code{occurrences},
\code{predictions} or \code{models}.}

\item{path}{A path with filename and the proper extension (see details) or the directory to save
files in.}

\item{file_path}{A path to save the \code{sdm_area} GeoPackage file.}

\item{file_name}{The name of the \code{sdm_area} GeoPackage file to be saved without extension.}

\item{extension}{How it should be saved?}
}
\description{
This function exports caretSDM data.
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
