% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_ensembles}
\alias{write_ensembles}
\title{Write/read caretSDM data into R.}
\usage{
write_ensembles(x, path = "results/ensembles", ext = ".tif")
}
\arguments{
\item{x}{Object to be written}

\item{path}{A path with filename and the proper extension (see details) or the directory to save files in.}

\item{extension}{How it should be saved?}

\item{file}{File to be imported with proper path when needed.}
}
\description{
This function saves and loads caretSDM data.
}
\details{
...
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
